+++ doc/release/release_notes.rst
@@ -7,10 +7,11 @@
 Release Notes
 *************
 
-.. release:: Upcoming
+.. release:: 1.0.0
+    :date: 2017-07-07
     
     .. change:: fixed
-        :tag: Logging
+        :tags: Logging
 
         Legacy api event hub spams Maya.